+++ src/rt/access.rs
@@ -3,6 +3,7 @@
 //! [`rt::Access`]: crate::rt::Access
 
 use std::ops::{Deref, DerefMut};
+use std::sync::Arc;
 use std::time::Instant;
 use std::{fmt, io};
 
@@ -11,7 +12,7 @@ use mio::{event, Interest, Token};
 use crate::actor::{self, AddActorError, NewActor, PrivateSpawn, Spawn};
 use crate::actor_ref::ActorRef;
 use crate::rt::process::ProcessId;
-use crate::rt::{ActorOptions, RuntimeRef};
+use crate::rt::{shared, ActorOptions, RuntimeRef};
 use crate::supervisor::Supervisor;
 
 /// Trait to indicate an API needs access to the Heph runtime.
@@ -133,5 +134,89 @@ impl<S, NA> PrivateSpawn<S, NA, ThreadLocal> for ThreadLocal {
 impl fmt::Debug for ThreadLocal {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("ThreadLocal")
+    }
+}
+
+/// Provides access to the thread-safe parts of the runtime.
+///
+/// This implements the [`Access`] trait, which is required by various APIs to
+/// get access to the runtime.
+///
+/// This is usually a part of the [`actor::Context`], see it for more
+/// information.
+///
+/// [`actor::Context`]: crate::actor::Context
+pub struct ThreadSafe {
+    rt: Arc<shared::RuntimeInternals>,
+}
+
+impl ThreadSafe {
+    pub(crate) const fn new(rt: Arc<shared::RuntimeInternals>) -> ThreadSafe {
+        ThreadSafe { rt }
+    }
+}
+
+impl Access for ThreadSafe {}
+
+impl PrivateAccess for ThreadSafe {
+    fn register<S>(&mut self, source: &mut S, token: Token, interest: Interest) -> io::Result<()>
+    where
+        S: event::Source + ?Sized,
+    {
+        self.rt.register(source, token, interest)
+    }
+
+    fn reregister<S>(&mut self, source: &mut S, token: Token, interest: Interest) -> io::Result<()>
+    where
+        S: event::Source + ?Sized,
+    {
+        self.rt.reregister(source, token, interest)
+    }
+
+    fn add_deadline(&mut self, pid: ProcessId, deadline: Instant) {
+        self.rt.add_deadline(pid, deadline)
+    }
+
+    fn cpu(&self) -> Option<usize> {
+        None
+    }
+}
+
+impl<S, NA> Spawn<S, NA, ThreadSafe> for ThreadSafe
+where
+    S: Send + Sync,
+    NA: NewActor<RuntimeAccess = ThreadSafe> + Send + Sync,
+    NA::Actor: Send + Sync,
+    NA::Message: Send,
+{
+}
+
+impl<S, NA> PrivateSpawn<S, NA, ThreadSafe> for ThreadSafe
+where
+    S: Send + Sync,
+    NA: NewActor<RuntimeAccess = ThreadSafe> + Send + Sync,
+    NA::Actor: Send + Sync,
+    NA::Message: Send,
+{
+    fn try_spawn_setup<ArgFn, ArgFnE>(
+        &mut self,
+        supervisor: S,
+        new_actor: NA,
+        arg_fn: ArgFn,
+        options: ActorOptions,
+    ) -> Result<ActorRef<NA::Message>, AddActorError<NA::Error, ArgFnE>>
+    where
+        S: Supervisor<NA> + 'static,
+        NA: NewActor<RuntimeAccess = ThreadSafe> + 'static,
+        NA::Actor: 'static,
+        ArgFn: FnOnce(&mut actor::Context<NA::Message, ThreadSafe>) -> Result<NA::Argument, ArgFnE>,
+    {
+        self.rt.spawn_setup(supervisor, new_actor, arg_fn, options)
+    }
+}
+
+impl fmt::Debug for ThreadSafe {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str("ThreadSafe")
     }
 }
