+++ tests/functional/actor_context.rs
@@ -5,8 +5,8 @@
 use std::pin::Pin;
 use std::task::Poll;
 
-use heph::actor::{self, NoMessages, RecvError, ThreadSafe};
-use heph::rt::{ActorOptions, Runtime, Signal};
+use heph::actor::{self, NoMessages, RecvError};
+use heph::rt::{ActorOptions, Runtime, Signal, ThreadSafe};
 use heph::supervisor::NoSupervisor;
 use heph::test::{init_local_actor, poll_actor};
