+++ tests/functional/restart_supervisor.rs
@@ -10,8 +10,8 @@ use std::task::{self, Poll};
 use std::thread::sleep;
 use std::time::Duration;
 
-use heph::actor::{self, ThreadSafe};
-use heph::{restart_supervisor, Actor, NewActor, Supervisor, SupervisorStrategy};
+use heph::rt::ThreadSafe;
+use heph::{actor, restart_supervisor, Actor, NewActor, Supervisor, SupervisorStrategy};
 
 // NOTE: keep in sync with the documentation.
 const DEFAULT_MAX_RESTARTS: usize = 5;
